requirements.txt
pdfplumber==0.10.3
PyMuPDF==1.23.8
psycopg2-binary==2.9.9
qdrant-client==1.7.0
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.24.0
apache-airflow==2.7.3
python-dotenv==1.0.0
